rename files and reorganise
