+++ app/processing/invoice-lines/validate-invoice-lines.js
@@ -1,4 +1,4 @@
-const schema = require('./invoice-line-schema')
+const schema = require('./schema')
 
 const validateInvoiceLines = (invoiceLines, paymentRequestId) => {
   invoiceLines.forEach(invoiceLine => {
